requirements.txt
streamlit
pandas
openpyxl
python-docx
pymupdf
firebase-admin
spacy==3.7.4
numpy==1.26.4
https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.8.0/pl_core_news_sm-3.8.0-py3-none-any.whl